move StorageNodeActorProtocol classes to justin.db.actors.protocol package
