+++ justin-core/src/main/scala/justin/db/actors/StorageNodeActor.scala
@@ -1,14 +1,11 @@
 package justin.db.actors
 
-import java.util.UUID
-
 import akka.actor.{Actor, ActorRef, Props}
 import justin.consistent_hashing.{NodeId, Ring}
 import justin.db._
-import justin.db.actors.StorageNodeActorProtocol.{StorageNodeReadData, StorageNodeWriteData}
-import justin.db.actors.protocol.{ReadData, WriteData}
-import justin.db.replica._
 import justin.db.storage.PluggableStorageProtocol
+import justin.db.actors.protocol.{ReadData, StorageNodeReadData, StorageNodeWriteData, WriteData}
+import justin.db.replica._
 
 class StorageNodeActor(nodeId: NodeId, storage: PluggableStorageProtocol, ring: Ring, n: N) extends Actor with ClusterSubscriberActor {
 
@@ -42,38 +39,3 @@ object StorageNodeActor {
 
 case class StorageNodeActorRef(storageNodeActor: ActorRef) extends AnyVal
 
-object StorageNodeActorProtocol {
-
-  // ----- READ PART ----
-  // INPUT
-  sealed trait StorageNodeReadData
-  object StorageNodeReadData {
-    case class Local(id: UUID)            extends StorageNodeReadData
-    case class Replicated(r: R, id: UUID) extends StorageNodeReadData
-  }
-  // OUTPUT
-  sealed trait StorageNodeReadingResult
-  object StorageNodeReadingResult {
-    case class Found(data: Data)           extends StorageNodeReadingResult
-    case class Conflicts(data: List[Data]) extends StorageNodeReadingResult
-    case object NotFound                   extends StorageNodeReadingResult
-    case object FailedRead                 extends StorageNodeReadingResult
-  }
-  // ------
-
-  // ----- WRITE PART ----
-  // INPUT
-  sealed trait StorageNodeWriteData
-  object StorageNodeWriteData {
-    case class Local(data: Data)           extends StorageNodeWriteData
-    case class Replicate(w: W, data: Data) extends StorageNodeWriteData
-  }
-  // OUTPUT
-  sealed trait StorageNodeWritingResult
-  object StorageNodeWritingResult {
-    case object SuccessfulWrite                              extends StorageNodeWritingResult
-    case object FailedWrite                                  extends StorageNodeWritingResult
-    case class ConflictedWrite(oldData: Data, newData: Data) extends StorageNodeWritingResult
-  }
-  // ------
-}
