+++ justin-core/src/main/scala/justin/db/actors/protocol/ReplicaCoordinatorActorProtocol.scala
@@ -2,7 +2,6 @@ package justin.db.actors.protocol
 
 import akka.actor.ActorRef
 import justin.db.ClusterMembers
-import justin.db.actors.StorageNodeActorProtocol.{StorageNodeReadData, StorageNodeWriteData}
 
 case class WriteData(sender: ActorRef, clusterMembers: ClusterMembers, cmd: StorageNodeWriteData)
 case class ReadData(sender: ActorRef, clusterMembers: ClusterMembers, cmd: StorageNodeReadData)
