+++ justin-core/src/main/scala/justin/db/client/ActorRefStorageNodeClient.scala
@@ -6,7 +6,7 @@ import akka.pattern.ask
 import akka.util.Timeout
 import justin.db.replica.W
 import justin.db.Data
-import justin.db.actors.StorageNodeActorProtocol.{StorageNodeReadData, StorageNodeReadingResult, StorageNodeWriteData, StorageNodeWritingResult}
+import justin.db.actors.protocol.{StorageNodeReadData, StorageNodeReadingResult, StorageNodeWriteData, StorageNodeWritingResult}
 import justin.db.actors.StorageNodeActorRef
 import justin.db.replica.{R, W}
