+++ justin-core/src/main/scala/justin/db/replica/ReplicaRemoteWriter.scala
@@ -3,7 +3,7 @@ package justin.db.replica
 import akka.pattern.ask
 import akka.util.Timeout
 import justin.db.Data
-import justin.db.actors.StorageNodeActorProtocol.{StorageNodeWriteData, StorageNodeWritingResult}
+import justin.db.actors.protocol.{StorageNodeWriteData, StorageNodeWritingResult}
 import justin.db.actors.StorageNodeActorRef
 
 import scala.concurrent.duration._
