+++ justin-core/src/test/scala/justin/db/replica/ReplicaLocalWriterTest.scala
@@ -4,8 +4,8 @@ import java.util.UUID
 
 import justin.consistent_hashing.NodeId
 import justin.db.Data
-import justin.db.actors.StorageNodeActorProtocol.StorageNodeWritingResult
 import justin.db.replica.ReplicaLocalWriter
+import justin.db.actors.protocol.StorageNodeWritingResult
 import justin.db.storage.PluggableStorageProtocol.{Ack, DataOriginality, StorageGetData, StoragePutData}
 import justin.db.storage.{GetStorageProtocol, PutStorageProtocol}
 import justin.vector_clocks.{Counter, VectorClock}
