+++ justin-core/src/test/scala/justin/db/replica/ReplicaReadAgreementTest.scala
@@ -3,8 +3,9 @@ package justin.db.replica
 import java.util.UUID
 
 import justin.db.Data
-import justin.db.actors.StorageNodeActorProtocol.StorageNodeReadingResult.{FailedRead, Found, NotFound}
 import justin.db.replica.ReplicaReadAgreement.ReadAgreement
+import justin.db.actors.protocol.StorageNodeReadingResult.{FailedRead, Found, NotFound}
+import org.scalatest.{FlatSpec, Matchers}
 import justin.db.versioning.VectorClockOps._
 import org.scalatest.{FlatSpec, Matchers}
