+++ justin-core/src/test/scala/justin/db/replica/ReplicaRemoteReaderTest.scala
@@ -5,7 +5,7 @@ import java.util.UUID
 import akka.actor.{Actor, ActorSystem}
 import akka.testkit.{TestActorRef, TestKit}
 import justin.db.Data
-import justin.db.actors.StorageNodeActorProtocol.{StorageNodeReadData, StorageNodeReadingResult}
+import justin.db.actors.protocol.{StorageNodeReadData, StorageNodeReadingResult}
 import justin.db.actors.StorageNodeActorRef
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{FlatSpecLike, Matchers}
