+++ justin-core/src/test/scala/justin/db/replica/ReplicaWriteAgreementTest.scala
@@ -1,8 +1,7 @@
 package justin.db.replica
 
-import justin.db.actors.StorageNodeActorProtocol.StorageNodeWritingResult.{FailedWrite, SuccessfulWrite}
 import justin.db.replica.ReplicaWriteAgreement.WriteAgreement
-import justin.db.replica.{ReplicaWriteAgreement, W}
+import justin.db.actors.protocol.StorageNodeWritingResult.{FailedWrite, SuccessfulWrite}
 import org.scalatest.{FlatSpec, Matchers}
 
 class ReplicaWriteAgreementTest extends FlatSpec with Matchers {
